Chore: move buttons to project list
