+++ frontend/src/component/project/ProjectList/ProjectList.tsx
@@ -16,7 +16,12 @@ import Add from '@mui/icons-material/Add';
 import ApiError from 'component/common/ApiError/ApiError';
 import useUiConfig from 'hooks/api/getters/useUiConfig/useUiConfig';
 import { TablePlaceholder } from 'component/common/Table';
-import { useMediaQuery, styled } from '@mui/material';
+import {
+    useMediaQuery,
+    styled,
+    ToggleButtonGroup,
+    ToggleButton,
+} from '@mui/material';
 import theme from 'themes/theme';
 import { Search } from 'component/common/Search/Search';
 import { PremiumFeature } from 'component/common/PremiumFeature/PremiumFeature';
@@ -25,6 +30,7 @@ import { ReactComponent as ProPlanIcon } from 'assets/icons/pro-enterprise-featu
 import { ReactComponent as ProPlanIconLight } from 'assets/icons/pro-enterprise-feature-badge-light.svg';
 import { safeRegExp } from '@server/util/escape-regex';
 import { ThemeMode } from 'component/common/ThemeMode/ThemeMode';
+import { useUiFlag } from 'hooks/useUiFlag';
 
 const StyledDivContainer = styled('div')(({ theme }) => ({
     display: 'flex',
@@ -47,6 +53,28 @@ const StyledCardLink = styled(Link)(({ theme }) => ({
     background: 'transparent',
     fontFamily: theme.typography.fontFamily,
     pointer: 'cursor',
+}));
+
+const StyledButtonGroup = styled(ToggleButtonGroup)(({ theme }) => ({
+    'html[data-theme="light"] button': {
+        '--hover-background-color': '#615BC2',
+    },
+    'html[data-theme="dark"] button': {
+        '--hover-background-color': '#34325E',
+    },
+
+    button: {
+        color: theme.palette.primary.main,
+        backgroundColor: theme.palette.background,
+        textTransform: 'capitalize',
+    },
+    'button[aria-pressed=true]': {
+        backgroundColor: theme.palette.background.alternative,
+        color: theme.palette.primary.contrastText,
+    },
+    'button[aria-pressed=true]:hover': {
+        backgroundColor: '#615BC2 /*var(--hover-background-color) */',
+    },
 }));
 
 type PageQueryType = Partial<Record<'search', string>>;
@@ -112,6 +140,10 @@ export const ProjectListNew = () => {
         searchParams.get('search') || '',
     );
 
+    const showProjectFilterButtons = useUiFlag('projectListFilterMyProjects');
+    const [filter, setFilter] = useState('all projects');
+    const filters = ['all projects', 'my projects'];
+
     useEffect(() => {
         const tableState: PageQueryType = {};
         if (searchValue) {
@@ -174,6 +206,33 @@ export const ProjectListNew = () => {
                     title={`Projects (${projectCount})`}
                     actions={
                         <>
+                            <ConditionallyRender
+                                condition={showProjectFilterButtons}
+                                show={
+                                    <StyledButtonGroup
+                                        aria-label='project list filter'
+                                        color='primary'
+                                        value={filter}
+                                        exclusive
+                                        onChange={(event, value) => {
+                                            if (value !== null) {
+                                                setFilter(value);
+                                            }
+                                        }}
+                                    >
+                                        {filters.map((filter) => {
+                                            return (
+                                                <ToggleButton
+                                                    key={filter}
+                                                    value={filter}
+                                                >
+                                                    {filter}
+                                                </ToggleButton>
+                                            );
+                                        })}
+                                    </StyledButtonGroup>
+                                }
+                            />
                             <ConditionallyRender
                                 condition={!isSmallScreen}
                                 show={
